requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cron==0.6.0
djangorestframework==3.15.1
idna==3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
